Nav Links
Not only is GitHub refusing to update the header image after unpublishing the site and changing the branch it is built from repeatedly, it is refusing to update the Nav bar, which builds and functions fine locally, and instead publishes only the word "branch," which exists exactly nowhere in my code. Excellent. What a fine utility.